+++ app/Models/User.php
@@ -22,7 +22,9 @@ class User extends Authenticatable
         'name',
         'email',
         'password',
-        'status'
+        'status',
+        'google_id',
+        'avatar',
     ];
 
     /**